+++ src/main/java/dao/MatHangDAO.java
@@ -3,19 +3,21 @@ package dao;
 import model.Category;
 import model.MatHang;
 
-import java.io.InputStream;
-import java.sql.*;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MatHangDAO extends DAO{
+public class MatHangDAO extends DAO {
 
-    public MatHangDAO () {
+    public MatHangDAO() {
         super();
     }
+
     List<Category> categories = null;
 
-    public MatHang getMatHangById(int matHangId){
+    public MatHang getMatHangById(int matHangId) {
 //        m.matHangID,m.matHangCode,m.name,m.image,
 //m.retailPrice,m.wholesalePrice,m.description,
 //m.categoryID,m.createdDate,m.calculateUnit,
@@ -26,21 +28,21 @@ public class MatHangDAO extends DAO{
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setInt(1, matHangId);
             ResultSet rs = ps.executeQuery();
-            if(rs.next()) {
-                result.setId(Long.parseLong(matHangId+""));
+            if (rs.next()) {
+                result.setId(Long.parseLong(matHangId + ""));
                 result.setCode(rs.getString("matHangCode"));
                 result.setName(rs.getString("name"));
                 result.setImage(rs.getBlob("image").getBinaryStream());
-                result.setRetailPrice(rs.getDouble("retailPrice"));
-                result.setWholesalePrice(rs.getDouble("wholesalePrice"));
+                result.setRetailPrice(rs.getLong("retailPrice"));
+                result.setWholesalePrice(rs.getLong("wholesalePrice"));
                 result.setDescription(rs.getString("description"));
                 int categoryId = rs.getInt("categoryID");
                 Category category = new Category();
                 CategoryDAO categoryDAO = new CategoryDAO();
                 categories = categoryDAO.getListCategory();
 
-                for(int i = 0; i < categories.size(); i++) {
-                    if(categories.get(i).getCategoryId() == categoryId){
+                for (int i = 0; i < categories.size(); i++) {
+                    if (categories.get(i).getCategoryId() == categoryId) {
                         category = categories.get(i);
                     }
                 }
@@ -51,7 +53,7 @@ public class MatHangDAO extends DAO{
                 result.setWeight(rs.getFloat("weight"));
                 result.setAttribute(rs.getString("attribute"));
             }
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
         }
         System.out.println(result.getName());
@@ -80,7 +82,7 @@ public class MatHangDAO extends DAO{
             ps.setInt(11, matHang.getUnit());
             ps.setFloat(12, matHang.getWeight());
             ps.executeQuery();
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             check = false;
         }
@@ -96,7 +98,7 @@ public class MatHangDAO extends DAO{
         String sql = "call updateMatHangWithoutImage(?,?,?,?,?,?,?,?,?,?,?)";
         try {
             PreparedStatement ps = con.prepareStatement(sql);
-            ps.setInt(1, Integer.parseInt(matHang.getId()+""));
+            ps.setInt(1, Integer.parseInt(matHang.getId() + ""));
             ps.setString(2, matHang.getCode());
             ps.setString(3, matHang.getName());
 //            ps.setBlob(4, matHang.getImage());
@@ -110,12 +112,13 @@ public class MatHangDAO extends DAO{
             ps.setFloat(11, matHang.getWeight());
             ps.executeQuery();
             System.out.println("Done update");
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             check = false;
         }
         return check;
     }
+
     public boolean updateMatHangWithImage(MatHang matHang) {
         //matHangID, matHangCode , name,
         // image, retailPrice,
@@ -125,7 +128,7 @@ public class MatHangDAO extends DAO{
         String sql = "call updateMatHangWithImage(?,?,?,?,?,?,?,?,?,?,?,?)";
         try {
             PreparedStatement ps = con.prepareStatement(sql);
-            ps.setInt(1, Integer.parseInt(matHang.getId()+""));
+            ps.setInt(1, Integer.parseInt(matHang.getId() + ""));
             ps.setString(2, matHang.getCode());
             ps.setString(3, matHang.getName());
             ps.setBlob(4, matHang.getImage());
@@ -139,14 +142,14 @@ public class MatHangDAO extends DAO{
             ps.setFloat(12, matHang.getWeight());
             ps.executeQuery();
             System.out.println("Done update");
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             check = false;
         }
         return check;
     }
 
-    public List<MatHang> getListMatHang(){
+    public List<MatHang> getListMatHang() {
         List<MatHang> arrayList = new ArrayList<>();
         String sql = "call getListMatHang(?)";
         MatHang mh = new MatHang();
@@ -154,11 +157,11 @@ public class MatHangDAO extends DAO{
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setInt(1, 1);
             ResultSet rs = ps.executeQuery();
-            if(rs.next()) {
+            if (rs.next()) {
                 mh.setName(rs.getString("name"));
                 arrayList.add(mh);
             }
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return arrayList;
@@ -183,59 +186,59 @@ public class MatHangDAO extends DAO{
         return list;
     }
 
-    public List<MatHang> getall(){
+    public List<MatHang> getall() {
         List<MatHang> list = new ArrayList<>();
-        String sql="call getListMatHang()";
-        try{
+        String sql = "call getListMatHang()";
+        try {
             PreparedStatement st = con.prepareStatement(sql);
             ResultSet rs = st.executeQuery();
-            while(rs.next()){
+            while (rs.next()) {
                 MatHang m = new MatHang();
                 m.setId(rs.getLong("matHangID"));
                 m.setName(rs.getString("name"));
                 m.setCode(rs.getString("matHangCode"));
-                m.setRetailPrice(rs.getDouble("retailPrice"));
-                m.setWholesalePrice(rs.getDouble("wholesalePrice"));
+                m.setRetailPrice(rs.getLong("retailPrice"));
+                m.setWholesalePrice(rs.getLong("wholesalePrice"));
                 m.setQuantity(rs.getInt("quantity"));
                 list.add(m);
             }
-        }catch (SQLException e){
+        } catch (SQLException e) {
             System.out.println(e);
         }
         return list;
     }
 
-    public void deleteMH(int id){
+    public void deleteMH(int id) {
 
-        try{
+        try {
             String sql = "call deleteMatHang(?)";
             PreparedStatement st = con.prepareStatement(sql);
-            st.setInt(1,id);
+            st.setInt(1, id);
             st.executeUpdate();
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
-    public List<MatHang> searchByName(String txtSearch){
+    public List<MatHang> searchByName(String txtSearch) {
         List<MatHang> list = new ArrayList<>();
         String sql = "call searchMatHang(?)";
 
-        try{
+        try {
             PreparedStatement st = con.prepareStatement(sql);
-            st.setString(1,"%"+txtSearch+"%");
+            st.setString(1, "%" + txtSearch + "%");
             ResultSet rs = st.executeQuery();
-            while(rs.next()){
+            while (rs.next()) {
                 MatHang m = new MatHang();
                 m.setId(rs.getLong("matHangID"));
                 m.setName(rs.getString("name"));
                 m.setCode(rs.getString("matHangCode"));
-                m.setRetailPrice(rs.getDouble("retailPrice"));
-                m.setWholesalePrice(rs.getDouble("wholesalePrice"));
+                m.setRetailPrice(rs.getLong("retailPrice"));
+                m.setWholesalePrice(rs.getLong("wholesalePrice"));
                 m.setCreatedDate(rs.getDate("createdDate"));
                 list.add(m);
             }
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return list;
